added more index creation logic
the flow of the code:
keeps a lexicon entry for current word for the blockid and blockoffset where it started,
and iterates through the postings, adding docids and corresponding frequencies to
separate memory blocks. When those blocks are full (should be at the same time), they're
appended to the large memory buffer of blocks. When that gets full, it gets compressed and piped out
to the inverted_index file.